Moving PRO3p1_decay_invis into its own files
